+++ store/rootmulti/store_test.go
@@ -87,7 +87,7 @@ func TestCacheMultiStoreWithVersion(t *testing.T) {
 	cID := ms.Commit()
 	require.Equal(t, int64(1), cID.Version)
 
-	// require no failure when given an invalid or pruned version
+	// require error when given an invalid or pruned version
 	_, err = ms.CacheMultiStoreWithVersion(cID.Version + 1)
 	require.Error(t, err)
 
@@ -535,11 +535,11 @@ func TestMultiStore_Pruning(t *testing.T) {
 		deleted     []int64
 		saved       []int64
 	}{
-		{"prune nothing", 10, pruningtypes.NewPruningOptions(pruningtypes.PruningNothing), nil, []int64{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}},
-		{"prune everything", 12, pruningtypes.NewPruningOptions(pruningtypes.PruningEverything), []int64{1, 2, 3, 4, 5, 6, 7}, []int64{8, 9, 10, 11, 12}},
-		{"prune some; no batch", 10, pruningtypes.NewCustomPruningOptions(2, 1), []int64{1, 2, 3, 4, 6, 5, 7}, []int64{8, 9, 10}},
-		{"prune some; small batch", 10, pruningtypes.NewCustomPruningOptions(2, 3), []int64{1, 2, 3, 4, 5, 6}, []int64{7, 8, 9, 10}},
-		{"prune some; large batch", 10, pruningtypes.NewCustomPruningOptions(2, 11), nil, []int64{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}},
+		{"prune nothing", 10, types.PruneNothing, nil, []int64{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}},
+		{"prune everything", 10, types.PruneEverything, []int64{1, 2, 3, 4, 5, 6, 7}, []int64{8, 9, 10}},
+		{"prune some; no batch", 10, types.NewPruningOptions(2, 3, 1), []int64{1, 2, 4, 5, 7}, []int64{3, 6, 8, 9, 10}},
+		{"prune some; small batch", 10, types.NewPruningOptions(2, 3, 3), []int64{1, 2, 4, 5}, []int64{3, 6, 7, 8, 9, 10}},
+		{"prune some; large batch", 10, types.NewPruningOptions(2, 3, 11), nil, []int64{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}},
 	}
 
 	for _, tc := range testCases {
@@ -552,18 +552,14 @@ func TestMultiStore_Pruning(t *testing.T) {
 				ms.Commit()
 			}
 
-			for _, v := range tc.deleted {
-				// Ensure async pruning is done
-				checkErr := func() bool {
-					_, err := ms.CacheMultiStoreWithVersion(v)
-					return err != nil
-				}
-				require.Eventually(t, checkErr, 1*time.Second, 10*time.Millisecond, "expected error when loading height: %d", v)
+			for _, v := range tc.saved {
+				_, err := ms.CacheMultiStoreWithVersion(v)
+				require.NoError(t, err, "expected no error when loading height: %d", v)
 			}
 
 			for _, v := range tc.saved {
 				_, err := ms.CacheMultiStoreWithVersion(v)
-				require.NoError(t, err, "expected no error when loading height: %d", v)
+				require.Error(t, err, "expected error when loading height: %d", v)
 			}
 		})
 	}
@@ -641,8 +637,11 @@ func TestMultiStore_PruningRestart(t *testing.T) {
 	// commit one more block and ensure the heights have been pruned
 	ms.Commit()
 
-	actualHeightToPrune = ms.pruningManager.GetPruningHeight(ms.LatestVersion())
-	require.Equal(t, int64(8), actualHeightToPrune)
+	for _, v := range pruneHeights {
+		_, err := ms.CacheMultiStoreWithVersion(v)
+		require.Error(t, err, "expected error when loading height: %d", v)
+	}
+}
 
 func TestSetInitialVersion(t *testing.T) {
 	db := coretesting.NewMemDB()